\name{aerobic_etc}
\alias{aerobic_etc}
\docType{data}
\title{
aerobic genes}
\description{
data set of aerobic related genes}
\usage{data("aerobic_etc")}
\format{
  A data frame with 87 observations each is an MGI symbol
}

\examples{
data(aerobic_etc)
## maybe str(aerobic_etc) ; plot(aerobic_etc) ...
}
\keyword{datasets}
