\name{allgenes}
\alias{allgenes}
\docType{data}
\title{
gene list}
\description{
list of all genes found in data set, for use in creating genemap}
\usage{data("allgenes")}
\format{
  A data frame with 55414 observations on the following variable.
  \describe{
    \item{\code{GeneID}}{a character vector}
  }
}

\examples{
data(allgenes)
## maybe str(allgenes) ; plot(allgenes) ...
}
\keyword{datasets}
