\name{anno_colors}
\alias{anno_colors}
\docType{data}
\title{
example annotation colors}
\description{
colors for use in heatmap annotation}
\usage{data("anno_colors")}
\format{
  The format is:
List of 1
 $ Group: Named chr [1:3] "blue" "green" "red"
  ..- attr(*, "names")= chr [1:3] "Naive" "Tolerant" "Rejecting"
}

\examples{
data(anno_colors)
## maybe str(anno_colors) ; plot(anno_colors) ...
}
\keyword{datasets}
