\name{anno_rlogdf}
\alias{anno_rlogdf}
\docType{data}
\title{
annotated rlog normalized data frame}
\description{
data frame containing rlog normalized data wiht annotations}
\usage{data("anno_rlogdf")}
\format{
  A data frame with 55414 observations on the following 17 variables.
  \describe{
    \item{\code{MGI_Symbol}}{a character vector}
    \item{\code{MGI_Desc}}{a character vector}
    \item{\code{Naive_11}}{a numeric vector}
    \item{\code{Naive_16}}{a numeric vector}
    \item{\code{Naive_38}}{a numeric vector}
    \item{\code{Naive_56}}{a numeric vector}
    \item{\code{Naive_82}}{a numeric vector}
    \item{\code{Naive_85}}{a numeric vector}
    \item{\code{Rejecting_74}}{a numeric vector}
    \item{\code{Rejecting_76}}{a numeric vector}
    \item{\code{Rejecting_78}}{a numeric vector}
    \item{\code{Rejecting_80}}{a numeric vector}
    \item{\code{Tolerant_34}}{a numeric vector}
    \item{\code{Tolerant_36}}{a numeric vector}
    \item{\code{Tolerant_68}}{a numeric vector}
    \item{\code{Tolerant_70}}{a numeric vector}
    \item{\code{Tolerant_72}}{a numeric vector}
  }
}


\examples{
data(anno_rlogdf)
## maybe str(anno_rlogdf) ; plot(anno_rlogdf) ...
}
\keyword{datasets}
