% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorynetplot.R
\name{categorynetplot}
\alias{categorynetplot}
\title{Create Category Netplot}
\usage{
categorynetplot(
  go_obj,
  deg_df,
  pval_cutoff = 0.05,
  l2fc_cutoff = 0.5,
  categories_to_show = 30,
  seed = 2,
  shape = "circle"
)
}
\arguments{
\item{go_obj}{go object, result of either getOE_analysis or get_GSEA}

\item{deg_df}{dataframe}

\item{pval_cutoff}{number}

\item{l2fc_cutoff}{number}

\item{categories_to_show}{number; defaults to 30}

\item{seed}{number; defaults to 2, for set seed for reproducibility}

\item{shape}{string; one of "star", "circle", "gem", "dh", 'graphopt', 'grid', 'mds', 'randomly', 'fr', 'kk', 'drl' or 'lgl'.}
}
\value{
category netplot
}
\description{
Create Category Netplot
}
\examples{
categorynetplot(goobj, one_vs_two)
}
