% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparativeVolcano.R
\name{comp_volcano}
\alias{comp_volcano}
\title{Make comparative volcano plot}
\usage{
comp_volcano(
  deg_df1,
  deg_df2,
  plotTitle,
  conditions = c("numerator1", "denominator1", "num2", "denom2"),
  l2fc_cutoff = 0.8,
  pval_cutoff = 0.45,
  dircolors = c("#a50000", "#00009c", "darkgreen", "purple4", "gray70"),
  dirshapes = c(24, 25, 23, 23, 21),
  overlaps = 50
)
}
\arguments{
\item{deg_df1}{dataframe of pairwise comparison results, must be annotated}

\item{deg_df2}{dataframe of pairwise comparison results, must be annotated}

\item{plotTitle}{string, plot title}

\item{conditions}{list of 4 strings, c("numerator for df1", "denominator for df1", "numerator for df2", "denominator for df2")}

\item{l2fc_cutoff}{number, where to draw the cutoff lines for log2 fold changes and to determine which points to color/label}

\item{pval_cutoff}{number, points with P value lower than this in either dataframe will not be plotted}

\item{dircolors}{list of 5 colors (strings) for both up, both down, up in 1 & down in 2, up in 2 & down in 1, and not changed}

\item{dirshapes}{list of 5 shapes (numbers) for both up, both down, up in 1 & down in 2, up in 2 & down in 1, and not changed}

\item{overlaps}{number, indicates max overlaps for the labeling}
}
\value{
plot of log2 fold change vs log 2 fold change
}
\description{
Make comparative volcano plot
}
\examples{
comp_volcano(Rej_vs_Tol, Naive_vs_Tol, "Naive and Rejecting comparison", "Rejecting", "Naive", "tolerant")
}
