% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDEGlists.R
\name{compareDEGlists}
\alias{compareDEGlists}
\title{Compare two lists of DEGs}
\usage{
compareDEGlists(
  deg_df1,
  deg_df2,
  conditions = c("numerator1", "denominator1", "num2", "denom2"),
  l2fc_cutoff = 0.8,
  pval_cutoff = 0.45,
  export = FALSE
)
}
\arguments{
\item{deg_df1}{dataframe of DEGs from a pairwise comparison}

\item{deg_df2}{dataframe of DEGs from a pairwise comparison}

\item{conditions}{list of 4 strings, c("numerator for df1", "denominator for df1", "numerator for df2", "denominator for df2")}

\item{l2fc_cutoff}{number, the cutoff for up or down based on log2 fold change}

\item{pval_cutoff}{pvalue cut off for which points to exclude}

\item{export}{should the results also be exported to CSVs}
}
\value{
dataframe of genes with l2fc, pval for both conditions and the mgi symbol and description
}
\description{
Compare two lists of DEGs
}
\examples{
compareDEGlists(one_vs_two, three_vs_two, conditions = c("one", "two", "three", "two"), l2fc_cutoff = 1.2, pval_cutoff = 0.25)
}
