\name{condition_labels}
\alias{condition_labels}
\docType{data}
\title{
Condition labels}
\description{
example of condition labels}
\usage{data("condition_labels")}
\format{
  The format is:
 Factor w/ 3 levels "Naive","Tolerant",..: 1 2 3
}

\examples{
data(condition_labels)
## maybe str(condition_labels) ; plot(condition_labels) ...
}
\keyword{datasets}
