% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidateGenelists.R
\name{consolidate_gene_list}
\alias{consolidate_gene_list}
\title{Consolidate Gene List CSVs}
\usage{
consolidate_gene_list(map, ...)
}
\arguments{
\item{map}{dataframe, genemap}

\item{...}{strings, file names of CSVs to consolidate}
}
\value{
data frame, all unique genes from each data frame combined
}
\description{
Consolidate Gene List CSVs
}
\examples{
consolidate_gene_list(gmap, "oxphos.csv", "activation.csv", "proliferation.csv")
}
