% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGenemap.R
\name{createGenemap}
\alias{createGenemap}
\title{Create Genemap}
\usage{
createGenemap(
  allcounts,
  bio_mart = "ensembl",
  map_genome = "mmusculus_gene_ensembl",
  GMattributes = c("ensembl_gene_id", "mgi_symbol", "mgi_description", "entrezgene_id",
    "gene_biotype")
)
}
\arguments{
\item{allcounts}{data frame that at a minimum includes ensembl Gene ID for all of the genes in the data set in a column called GeneID}

\item{bio_mart, }{string, defaults to ensembl, dictates what is used to build the genemap}

\item{map_genome}{string, defaults to "mmusculus_gene_ensembl", dictates which genome to use to build the gene map}

\item{GMattributes}{list of strings, attributes to include in genemap, defaults to ensmbl ID, mgi symbol, mgi description and gene biotype}
}
\value{
dataframe containing all ensmbl IDs and corresponding MGI symbols and descriptions and gene biotypes
}
\description{
create a reference dataframe containing ensembl gene IDs and corresponding MGI symbols for use in data annotation
}
\examples{
createGenemap(allcounts = allgenes)
createGenemap(allgenes, GMattributes = c("ensembl_gene_id", "mgi_symbol"))

}
