\name{dseqobj}
\alias{dseqobj}
\docType{data}
\title{
DE seq object}
\description{
example deseq object}
\usage{data("dseqobj")}
\format{
  The format is:
Formal class 'DESeqDataSet' [package "DESeq2"] with 8 slots
  ..@ design            :Class 'formula'  language ~Group

}

\source{
Alegre Lab}

\examples{
data(dseqobj)
## maybe str(dseqobj) ; plot(dseqobj) ...
}
\keyword{datasets}
