% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportDEGs.R
\name{export_genelist}
\alias{export_genelist}
\title{Export Gene List}
\usage{
export_genelist(tab, filename, direction = NULL)
}
\arguments{
\item{tab}{data frame, contains ranked DEGs, the output of signif_deg}

\item{filename}{string, name to save the exported file as, avoid whitespace}

\item{direction}{string ("up", "down", or NULL), which genes to export}
}
\value{
CSV file exported to CWD
}
\description{
Export Gene List
}
\examples{
export_genelist(ranked_naive_vs_tol, "Naive-vs-tol")
export_genelist(ranked_naive_vs_tol, "Naive-vs-tol", direction = "up")
export_genelist(ranked_naive_vs_tol, "Naive-vs-tol", direction = "down")
}
