% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDEseq.R
\name{getDESeq}
\alias{getDESeq}
\title{getDESeq}
\usage{
getDESeq(countsmatrix, metadata, expt_design)
}
\arguments{
\item{countsmatrix}{data frame, cleaned up raw counts dataframe from RNA seq experiment}

\item{metadata}{data frame, cleaned up metadata table from RNA seq experiment}

\item{expt_design}{formula, experimental design}
}
\value{
DESeqDataSet object
}
\description{
This function will return the DESeq data object from which normalized counts and pairwise comparison results can be generated
}
\examples{
getDESeq(raw, meta, ~Group)
}
