% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldchange.R
\name{getFoldChanges}
\alias{getFoldChanges}
\title{Get Fold Change List}
\usage{
getFoldChanges(degs)
}
\arguments{
\item{degs}{df of pairwise comparison genes, not filtered}
}
\value{
list of fold changes with names of gene ids
}
\description{
Get Fold Change List
}
\examples{
getFoldChanges(one_vs_two)
}
