% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGSEA.R
\name{getGSEAobj}
\alias{getGSEAobj}
\title{Get GSEA}
\usage{
getGSEAobj(degs, ontol = "ALL")
}
\arguments{
\item{degs}{annotated, pairwise comparison dataframe, *not* the significant genes}

\item{ontol}{string, either "BP", "MF", "CC", or "ALL", defaults to ALL}
}
\value{
GSEA object
}
\description{
Get GSEA
}
\examples{
getGSEAobj(tol_vs_naive, ontol = "BP")
}
