% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{getPalettes}
\alias{getPalettes}
\title{Palettes}
\usage{
getPalettes()
}
\value{
prints the options for colors
}
\description{
just sets the color palettes and prints the options
}
\examples{
getPalettes()
}
