\name{gmap}
\alias{gmap}
\docType{data}
\title{
Gene map}
\description{
example gene map}
\usage{data("gmap")}
\format{
  A data frame with 55281 observations on the following 4 variables.
  \describe{
    \item{\code{ensembl_gene_id}}{a character vector}
    \item{\code{mgi_symbol}}{a character vector}
    \item{\code{mgi_description}}{a character vector}
    \item{\samp{genemap$ensembl_gene_id}}{a character vector}
  }
}

\examples{
data(gmap)
## maybe str(gmap) ; plot(gmap) ...
}
\keyword{datasets}
