% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overexpressionGOanalysis.R
\name{go_ORA}
\alias{go_ORA}
\title{Over-expression GO Term Analysis}
\usage{
go_ORA(deg_df, pval_cutoff = 0.05, l2fc_cutoff = 0.5, ontol = "ALL")
}
\arguments{
\item{deg_df}{dataframe}

\item{pval_cutoff}{number}

\item{l2fc_cutoff}{number}

\item{ontol}{string; either "ALL", "BP", "MF", or "CC"}
}
\value{
enrichResult
}
\description{
Over-expression GO Term Analysis
}
\examples{
getOE_analysis(one_vs_two)
}
