% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goGrouping.R
\name{goGrouping}
\alias{goGrouping}
\title{Group Go
Get a list of GO terms at a designated level that are represented by a list of genes input}
\usage{
goGrouping(genes, pval_cutoff = 0.05, l2fc_cutoff = 0.5, ontol = "BP", lvl = 2)
}
\arguments{
\item{genes}{dataframe, list of DEGs, must be annotated and have Log2FoldChange and Adj_P_Value columns}

\item{pval_cutoff}{number, defaults to 0.05}

\item{l2fc_cutoff}{number, defaults to 0.5}

\item{ontol}{string, "BP", "CC", or "MF" only}

\item{lvl}{number, defaults to 2, the level of GO to query}
}
\value{
gg, dataframe
}
\description{
Group Go
Get a list of GO terms at a designated level that are represented by a list of genes input
}
\examples{
goGrouping(one_vs_two, 0.01, 0.8, "BP", 4)
}
