% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goGrouping.R
\name{goGrouping_all_levels}
\alias{goGrouping_all_levels}
\title{Get Go Groups from all levels (2-6)}
\usage{
goGrouping_all_levels(
  rlogdf,
  pval_cutoff = 0.05,
  l2fc_cutoff = 0.5,
  ontol = c("BP", "MF", "CC", "ALL", "BP_MF")
)
}
\arguments{
\item{rlogdf}{rlog dataframe - ideally filtered to only include significantly changed/relevant genes}

\item{pval_cutoff}{number, defaults to 0.05}

\item{l2fc_cutoff}{number, defaults to 0.5}

\item{ontol}{string, "BP", "CC",  "MF", "ALL", or "BP_MF"}
}
\value{
gg, dataframe
}
\description{
Get Go Groups from all levels (2-6)
}
\examples{
goGrouping_all_levels(one_vs_two, 0.01, 0.8, "BP")
}
