% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapFromCSV.R
\name{heatmapFromCSV}
\alias{heatmapFromCSV}
\title{Heatmap from CSV}
\usage{
heatmapFromCSV(
  plottitle,
  fname,
  met,
  counts_df,
  annocolors,
  clusterRows = TRUE,
  gaps = c(),
  height = 10,
  colorscale = c("blue4", "white", "red3"),
  scaling = TRUE,
  displNumbers = FALSE
)
}
\arguments{
\item{plottitle}{string, title of the plot}

\item{fname}{string, full filename for the csv listing genes of interest, including path and .csv}

\item{met}{data frame, meta data information}

\item{counts_df}{data frame, normalized counts}

\item{annocolors}{list of strings, list of colors for use in labeling}

\item{clusterRows}{boolean, defaults to TRUE, should the rows cluster by similarity}

\item{gaps}{list of numbers, defaults to empty list, dictates where there would be gaps if desired between conditions}

\item{height}{number, defaults to 10, dictates how tall the cells in the heatmap will be}

\item{colorscale}{list of 3 colors sets the color scale, defaults to red (hi) white blue (low)}

\item{scaling}{boolean, defaults to TRUE - if rows should be scaled or not}

\item{displNumbers}{boolean, defaults to FALSE - if the value of each cell should be displayed within}
}
\value{
heatmap plot
}
\description{
Heatmap from CSV
}
\examples{
heatmapFromCSV("OxPhos", "oxphos.csv", metadata, rlog_df, c("blue", "red"))
}
