% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapFromDF.R
\name{heatmapFromGoGrouping}
\alias{heatmapFromGoGrouping}
\title{Heatmap from Go Grouping results}
\usage{
heatmapFromGoGrouping(
  counts_df,
  metadata,
  gogroup,
  anno_colors,
  rownum = NULL,
  ...
)
}
\arguments{
\item{counts_df}{dataframe containing annotated normalized counts}

\item{metadata}{dataframe with metadata info}

\item{gogroup}{dataframe with the output of goGrouping}

\item{anno_colors}{list of strings, list of colors for use in labeling}

\item{rownum}{number, which row of the go grouping results to plot}

\item{...}{any other parameters for create_heatmap()}
}
\value{
heatmap made with pheatmap
}
\description{
Heatmap from Go Grouping results
}
\examples{
heatmapFromGoGrouping(rlog_df, meta, tol_vs_rej_gogrouping, list(tol_color, rej_color, naive_color), 5)
heatmapFromGoGrouping(rlog_df, meta, tol_vs_rej_gogrouping, list(tol_color, rej_color, naive_color), 5, height = 6, gaps = c(3, 6))
}
