\name{IL2_STAT5_pathway}
\alias{IL2_STAT5_pathway}
\docType{data}
\title{
IL2/STAT5}
\description{
IL 2 and Stat 5 related gene pathway}
\usage{data("IL2_STAT5_pathway")}
\format{
  A data frame with 198 observations each is an MGI symbol

  }

\examples{
data(IL2_STAT5_pathway)
## maybe str(IL2_STAT5_pathway) ; plot(IL2_STAT5_pathway) ...
}
\keyword{datasets}
