% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGSEA.R
\name{kegg_GSEA}
\alias{kegg_GSEA}
\title{KEGG GSEA}
\usage{
kegg_GSEA(degs)
}
\arguments{
\item{degs}{annotated, pairwise comparison dataframe, *not* the significant genes}
}
\value{
GSEA object
}
\description{
KEGG GSEA
}
\examples{
kegg_GSEA(tol_vs_naive)
}
