% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overexpressionGOanalysis.R
\name{kegg_ORA}
\alias{kegg_ORA}
\title{Kegg over representation analysis}
\usage{
kegg_ORA(deg_df, pval_cutoff = 0.05, l2fc_cutoff = 0.5)
}
\arguments{
\item{deg_df}{dataframe}

\item{pval_cutoff}{number}

\item{l2fc_cutoff}{number}
}
\value{
enrichResult
}
\description{
Kegg over representation analysis
}
\examples{
kegg_ORA(one_vs_two)
}
