% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGOplot.R
\name{makeGOplot}
\alias{makeGOplot}
\title{Make GO Plot}
\usage{
makeGOplot(
  GO_obj,
  plot_title,
  plot_type,
  categories_to_show = 30,
  fontsize = 9
)
}
\arguments{
\item{GO_obj}{go object, result of either getOE_analysis or get_GSEA}

\item{plot_title}{string}

\item{plot_type}{string, one of "dot", "ridge", "cluster",}

\item{categories_to_show}{number; defaults to 30}

\item{fontsize}{number; defaults to 9}

\item{splitby}{string; defaults to NULL}
}
\value{
a plot or a string
}
\description{
Make GO Plot
}
\examples{
makeGOplot(onevstwo_GSEA_BP, "One Vs Two Dotplot", "dot")
}
