% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaDotplot.R
\name{makeGSEAdotplot}
\alias{makeGSEAdotplot}
\title{GSEA Dotplot}
\usage{
makeGSEAdotplot(
  gsea_obj,
  plottitle,
  categories_to_show = 25,
  font_size = 8,
  split = FALSE
)
}
\arguments{
\item{gsea_obj}{gene set object}

\item{plottitle}{string, plot title}

\item{categories_to_show}{number, defaults to 25}

\item{font_size}{number, defaults to 8}

\item{split}{boolean, defaults to FALSE}
}
\value{
dotplot
}
\description{
GSEA Dotplot
}
\examples{
makeGSEAdotplot(tolvsnaiveGSEA, "")
}
