% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaRidgeplot.R
\name{makeGSEAridgeplot}
\alias{makeGSEAridgeplot}
\title{GSEA ridgeplot}
\usage{
makeGSEAridgeplot(gsea_obj, plottitle, cats_to_show = 25, font_size = 8)
}
\arguments{
\item{gsea_obj}{gene set object}

\item{plottitle}{string}

\item{cats_to_show}{number}

\item{font_size}{number}
}
\value{
ridgeplot
}
\description{
GSEA ridgeplot
}
\examples{
makeGSEAridgeplot(tolvsnaiveGSEA, "")
}
