% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGOplot.R
\name{makeGSEplot}
\alias{makeGSEplot}
\title{Make Gene Set Enrichment Plot}
\usage{
makeGSEplot(
  GSE_obj,
  plot_title,
  plot_type,
  categories_to_show = 30,
  fontsize = 9,
  label_length = 75,
  foldChangeList = NULL,
  legend_loc = "bottom",
  seed = 5
)
}
\arguments{
\item{GSE_obj}{go object, result of either go_GSEA}

\item{plot_title}{string}

\item{plot_type}{string, one of "dot", "ridge", "cluster", "heat", "upset","tree", "traditional"}

\item{categories_to_show}{number; defaults to 30}

\item{fontsize}{number; defaults to 9}

\item{label_length}{number; defaults to 75, how long to let a label be before wrapping}

\item{foldChangeList}{list, defaults to null, fold changes with names attached - needed for coloring by FC}

\item{legend_loc}{string; defaults to "bottom" where the legend should be placed}

\item{seed}{number; defaults to 5, for reproducibility}
}
\value{
a plot or a string
}
\description{
Make Gene Set Enrichment Plot
}
\examples{
makeGSEplot(onevstwo_GSEA_BP, "One Vs Two Dotplot", "dot")
}
