\name{meta}
\alias{meta}
\docType{data}
\title{
Metadata example
}
\description{
example for metadata table that corresponds to example raw data }
\usage{data("meta")}
\format{
  A data frame with 15 observations on the following variable.
  \describe{
    \item{\code{Group}}{a factor with levels \code{Naive} \code{Tolerant} \code{Rejecting}}
  }
}
\source{
from Alegre Lab data}

\examples{
data(meta)
## maybe str(meta) ; plot(meta) ...
}
\keyword{datasets}
