% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/originalPCA.R
\name{originalPCA}
\alias{originalPCA}
\title{Old PCA function
this is the original verison of the PCA plot function. it uses the plot PCA function from the DESeq2 package.
there are less customize options in this than the other PCA plot}
\usage{
originalPCA(
  dseq_transform,
  plot_aes,
  plot_title = "Principal Component Analysis",
  label_samples = FALSE
)
}
\arguments{
\item{dseq_transform}{deseq transform object, rlog_norm not rlog_df}

\item{plot_aes}{the dataframe containing your plot aesthetics}

\item{plot_title}{string, plot title}

\item{label_samples}{boolean, should each dot be labeled}
}
\value{
pca plot
}
\description{
Old PCA function
this is the original verison of the PCA plot function. it uses the plot PCA function from the DESeq2 package.
there are less customize options in this than the other PCA plot
}
\examples{
originalPCA(rlog_norm, plot_aes, "PCA")
}
