% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overexpressionGOanalysis.R
\name{overexpression_analysis}
\alias{overexpression_analysis}
\title{Over-expression GO Term Analysis}
\usage{
overexpression_analysis(
  deg_df,
  pval_cutoff = 0.05,
  l2fc_cutoff = 0.5,
  ontol = "ALL"
)
}
\arguments{
\item{deg_df}{dataframe}

\item{pval_cutoff}{number}

\item{l2fc_cutoff}{number}

\item{ontol}{string; either "ALL", "BP", "MF", or "CC"}
}
\value{
data frame
}
\description{
Over-expression GO Term Analysis
}
\examples{
overexpression_analysis(one_vs_two)
}
