\name{oxphos_pathway}
\alias{oxphos_pathway}
\docType{data}
\title{
oxphos}
\description{Oxidative Phosphorylation gene set }
\usage{data("oxphos_pathway")}
\format{
  A data frame with 194 observations. Each is the MGI Symbol for a gene
  }

\examples{
data(oxphos_pathway)
## maybe str(oxphos_pathway) ; plot(oxphos_pathway) ...
}
\keyword{datasets}
