% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAplot.R
\name{pca_analysis}
\alias{pca_analysis}
\title{PCA Analysis}
\usage{
pca_analysis(
  dseq_transform,
  plot_aes,
  grouping_var = "Group",
  plot_title = "Principal Component Analysis",
  label_samples = FALSE
)
}
\arguments{
\item{dseq_transform}{DESeqTransform object}

\item{plot_aes}{dataframe, containing labels, colors, shapes, and fills for the data points being plotted}

\item{grouping_var}{string, defaults to Group, what is the grouping variable for this data}

\item{plot_title}{string, defaults to "Principal Component Analysis"}

\item{label_samples}{boolean, if true, labels for each sample will be added, defaults to FALSE}
}
\value{
PCA plot
}
\description{
Illustrate PCA plot to show sample relationships
}
\examples{
pca_analysis(rlognorm, aesthetics)
}
