% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneCounts.R
\name{plot_genecounts}
\alias{plot_genecounts}
\title{Plot Gene Counts}
\usage{
plot_genecounts(
  mgi,
  norm_df,
  metadat,
  plot_aes,
  comparisons = list(),
  grouping_var = "Group",
  label_samples = FALSE,
  overlaps_allowed = 10,
  yaxistozero = FALSE,
  kw_hjust = 0.5,
  kw_vjust = 1,
  yaxis = "rLog Normalized Reads"
)
}
\arguments{
\item{mgi}{string, MGI symbol for the gene of interest}

\item{norm_df}{data frame, transformed DESeq object as a data frame}

\item{metadat}{dataframe, meta data}

\item{plot_aes}{dataframe, containing labels, colors, shapes, and fills for the data points being plotted}

\item{comparisons}{list of length 2 lists, each length 2 list is the comparisons to use for stats, defaults to an empty list}

\item{grouping_var}{string, defaults to Group, what is the grouping variable for this data}

\item{label_samples}{boolean, defaults to FALSE, if true, adds labels for samples while avoiding overlaps}

\item{overlaps_allowed}{number, defaults to 10, increase to allow more overlaps in labeling}

\item{yaxistozero}{boolean, defaults to FALSE, if true the Y axis will}

\item{kw_hjust}{integer, dictates the horizontal position of the kruskal wallis value}

\item{kw_vjust}{integer, dictates the vertical position of the kruskal wallis value}

\item{yaxis}{string, label for the Y axis}
}
\value{
plot of normalized gene counts
}
\description{
Plot Gene Counts
}
\examples{
plot_genecounts("Tcf7", vst_df, metadata, treatment_aes, comparisons = stat_comparisons)

}
