% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printGroupNames.R
\name{print_group_names}
\alias{print_group_names}
\title{Print Group Names
Prints out group names. Pairwise comparison code must use group names to match exactly as returned here}
\usage{
print_group_names(dseqobj)
}
\arguments{
\item{dseqobj}{dseq object; the results of running the dseq code}
}
\value{
prints strings of group names
}
\description{
Print Group Names
Prints out group names. Pairwise comparison code must use group names to match exactly as returned here
}
\examples{
print_group_names(dseq)
}
