% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorderColumns.R
\name{reorder_columns}
\alias{reorder_columns}
\title{reorder Columns}
\usage{
reorder_columns(
  annotated_df,
  move2front = c("MGI_Symbol", "MGI_Desc", "GeneType")
)
}
\arguments{
\item{annotated_df}{dataframe, annotated results data frame, must include GeneID column}

\item{move2front}{list of strings, which columns to move to the front of the data frame, defaults to mgi symbol, description and genetype}
}
\value{
data frame with columns in correct order and geneID set to row name
}
\description{
reorder Columns
}
\examples{
reorder_columns(anno_tolvsnaive)
reorder_columns(anno_rejvstol, move2front = c("MGI_Symbol", "MGI_Symbol"))
}
