% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revigo.R
\name{revigoPlot}
\alias{revigoPlot}
\title{Revigo Plots
Creates a revigo plot}
\usage{
revigoPlot(
  gsea_obj,
  ontol = c("BP", "MF", "CC"),
  plotkind = c("tree", "scatter", "heat"),
  scoretype = c("pvalue", "qvalue", "nes"),
  plttitle = "",
  sim_threshold = 0.7
)
}
\arguments{
\item{gsea_obj}{gsea object - this must be generated with ALL ontology, if its pre-filitered the function wont work}

\item{ontol}{string, one of "BP", "MF", or "CC" - which ontology to use}

\item{plotkind}{string, one of "tree", "scatter", and "heat"; dictates plot type}

\item{scoretype}{string, one of "pvalue", "qvalue", "nes"; what is used for calculating the score for grouping/coloring blocks}

\item{plttitle}{string; plot title}

\item{sim_threshold}{number, similarity threshold Some guidance: Large (allowed similarity=0.9), Medium (0.7), Small (0.5), Tiny (0.4) Defaults to Medium (0.7)}
}
\value{
plot
}
\description{
Revigo Plots
Creates a revigo plot
}
\examples{
revigoPlot(gseTolvsRej, "BP", "tree", "pvalue")
}
