% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlog_filtering.R
\name{rlog_filtering}
\alias{rlog_filtering}
\title{Filter rLog DataFrame}
\usage{
rlog_filtering(p_thresh, fc_thresh, rlogdf, ...)
}
\arguments{
\item{p_thresh}{p value threshold, can be the same as your threshold for filtering pairwise comps but can be higher to preserve more genes}

\item{fc_thresh}{l2fc threshold, similar to above}

\item{rlogdf}{your annotated rlog_df}

\item{...}{any number of annotated, unfiltered, pairwise comparison results}
}
\value{
dataframe of rlog counts, filtered
}
\description{
This function will filter out genes that are not significantly different in any of the pairwise comparisons of interest.
This chunk will create a filtered dataframe only containing genes that reach significance in your specified pairwise comparisons.
This is helpful for removing genes that are not relevant to your conditions to minimize distracting noise.
}
\examples{
rlog_filtering(0.05, 0.5, rlog_df, one_vs_two, two_vs_three, three_vs_four)
}
