\name{rlogDF}
\alias{rlogDF}
\docType{data}
\title{
Rlog normalized data frame}
\description{
example data frame with rlog normalized data}
\usage{data("rlogDF")}
\format{
  A data frame with 55414 observations on the following 15 variables.
  \describe{
    \item{\code{Naive_11}}{a numeric vector}
    \item{\code{Naive_16}}{a numeric vector}
    \item{\code{Naive_38}}{a numeric vector}
    \item{\code{Naive_56}}{a numeric vector}
    \item{\code{Naive_82}}{a numeric vector}
    \item{\code{Naive_85}}{a numeric vector}
    \item{\code{Rejecting_74}}{a numeric vector}
    \item{\code{Rejecting_76}}{a numeric vector}
    \item{\code{Rejecting_78}}{a numeric vector}
    \item{\code{Rejecting_80}}{a numeric vector}
    \item{\code{Tolerant_34}}{a numeric vector}
    \item{\code{Tolerant_36}}{a numeric vector}
    \item{\code{Tolerant_68}}{a numeric vector}
    \item{\code{Tolerant_70}}{a numeric vector}
    \item{\code{Tolerant_72}}{a numeric vector}
  }
}

\source{
Alegre lab}

\examples{
data(rlogDF)
## maybe str(rlogDF) ; plot(rlogDF) ...
}
\keyword{datasets}
