\name{rlognorm}
\alias{rlognorm}
\docType{data}
\title{
R log normalized DE Seq transform object}
\description{
Example deseq transform object}
\usage{data("rlognorm")}
\format{
  The format is:
Formal class 'DESeqTransform' [package "DESeq2"] with 6 slots

}

\examples{
data(rlognorm)
## maybe str(rlognorm) ; plot(rlognorm) ...
}
\keyword{datasets}
