% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleDistances.R
\name{sample_distances}
\alias{sample_distances}
\title{Sample to Sample Distances}
\usage{
sample_distances(
  normobj,
  annotationcolors,
  metadat,
  clusterrows = FALSE,
  clustercols = FALSE,
  fontsize = 9
)
}
\arguments{
\item{normobj}{DESeqTransform object}

\item{annotationcolors}{list, list of colors for use in heat map}

\item{metadat}{data frame, meta-data data frame of 1 variable}

\item{clusterrows}{boolean, defaults to False, should there be row clustering in heatmap}

\item{clustercols}{boolean, defaults to False,should there be column clustering in heatmap}

\item{fontsize}{numeric, defaults to 9, font size for heatmap}
}
\value{
heatmap
}
\description{
Sample to Sample Distances
}
\examples{
sample_distances(rlog_norm,
                 list(Condition = c("Acute Rejection" = "#0404b9", Tolerant = "#389538")),
                 meta)
}
