% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_aes.R
\name{set_aes}
\alias{set_aes}
\title{Set Aesthetics}
\usage{
set_aes(label, shapes, outlines, fills)
}
\arguments{
\item{label}{list of strings}

\item{shapes}{list of strings}

\item{outlines}{list of strings}

\item{fills}{list of strings}
}
\value{
data frame, aesthetics for use in plotting
}
\description{
Set Aesthetics
}
\examples{
set_aes(
  condition_labels,
  c("circle", "square", "triangle up", "x"),
  c("blue", "green", "red", "orange"),
  c("blue", "green", "red", "orange"))

}
