% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMAPplot.R
\name{umap_analysis}
\alias{umap_analysis}
\title{UMAP Analysis}
\usage{
umap_analysis(
  dseq_transform,
  neighbors,
  m_dist,
  cond_list,
  plot_aes,
  plot_title = "UMAP",
  label_samples = FALSE,
  batch = FALSE
)
}
\arguments{
\item{dseq_transform}{DESeqTransform object}

\item{neighbors}{integer, minimum of 2}

\item{m_dist}{integer}

\item{cond_list}{list, conditions for each data point}

\item{plot_aes}{dataframe, containing labels, colors, shapes, and fills for the data points being plotted}

\item{plot_title}{string, plot title, defaults to "UMAP"}

\item{label_samples}{boolean, if sample labels should be added to plot, defaults to FALSE}

\item{batch}{boolean, defaults to false, set to true & add set seed for reproducible results}
}
\value{
UMAP plot
}
\description{
plot umap for all samples, reducing the number of variables and showing relationships between samples
}
\examples{
umap_analysis(rlognorm, 5, 10, plot_aes)
}
