% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennDiagram.R
\name{venn_overlaps}
\alias{venn_overlaps}
\title{Get Venn Diagram Overlaps}
\usage{
venn_overlaps(degs_list)
}
\arguments{
\item{degs_list}{list of named lists, lists of MGI symbols from significant DEGs of pairwise comparisons, should be named with the label}
}
\value{
dataframe of overlaps between comparisions
}
\description{
Get Venn Diagram Overlaps
}
\examples{
venn_overlaps(list("Treg" = ranked_TregNaive$MGI_Symbol, "Tol" = ranked_TolNaive$MGI_Symbol))
venn_overlaps(list("Treg" = ranked_TregNaive$MGI_Symbol, "Tol" = ranked_TolNaive$MGI_Symbol, "Rej" = ranked_RejNaive$MGI_Symbol))
}
