% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano plot}
\usage{
volcano_plot(
  deg_df,
  plotTitle,
  cond1,
  cond2,
  l2fc_cutoff = 2,
  pval_cutoff = 0.01,
  dircolors = c("#a50000", "#00009c", "gray70"),
  dirshapes = c(19, 19, 19)
)
}
\arguments{
\item{deg_df}{data frame, pairwise comparison results, should contain at minimum Log2FC MGI symbol, and adjusted p value}

\item{plotTitle}{string, plot title}

\item{cond1}{string, name of condition 1}

\item{cond2}{string, name of condition 2}

\item{l2fc_cutoff}{number, log2 fold change threshold for significance, defaults to 2}

\item{pval_cutoff}{number, adjusted p value threshold for significance, defaults to 0.01}

\item{dircolors}{list of strings, colors for significant up, down, and insignificant, defaults to red, blue and grey}

\item{dirshapes}{list of numbers, shapes for points up, down, and insignificant, defaults to circles for all}
}
\value{
ggplot volcano plot
}
\description{
creates a volcano plot showing the relationship between log2foldchange and p value for all genes in a pairwise comparison
}
\examples{
volcano_plot(tol_vs_naive, "Tolerant vs. Naive", "Tolerant", "Naive")
volcano_plot(tol_vs_naive, "Tolerant vs. Naive", "Tolerant", "Naive", pval_cutoff = 0.005)
}
