% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateResults.R
\name{annotate_biomart}
\alias{annotate_biomart}
\title{Annotate Biomart}
\usage{
annotate_biomart(res, gm, gmattri = c("MGI_Symbol", "MGI_Desc", "GeneType"))
}
\arguments{
\item{res}{dataframe built from DESeq results, must have ensembl gene ID in 'GeneID' column}

\item{gm}{biomart genemap, contains GeneID and MGI_symbol columns}
}
\value{
results object now with annotations
}
\description{
add human readable gene IDs, descriptions, and gene biotypes to the data
}
\examples{
annotate_biomart(tol_vs_naive, gmap)
annotate_biomart(tol_vs_naive, gmap, c("MGI_Symbol", "MGI_Desc"))
}
