% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{create_heatmap}
\alias{create_heatmap}
\title{Basic heatmap}
\usage{
create_heatmap(
  title,
  genes_and_counts,
  met,
  annocolors,
  clusterRows = TRUE,
  gaps = c(),
  height = 10,
  colorscale
)
}
\arguments{
\item{title}{string, title of the plot}

\item{genes_and_counts}{data frame, contains genes of interest and the corresponding normalized read counts}

\item{met}{data frame, meta data for the experiment}

\item{annocolors}{list of strings, list of colors for use in labeling}

\item{clusterRows}{boolean, defaults to TRUE, should the heatmap cluster rows by similarity}

\item{gaps}{list of numbers, defaults to empty list, dictates where there would be gaps if desired between conditions}

\item{height}{number, defaults to 10, dictates how tall the cells in the heatmap will be}

\item{colorscale}{list of 3 colors sets the color scale}
}
\value{
pheatmap plot
}
\description{
Basic heatmap
}
\examples{
create_heatmap("Oxphos Genes", oxhposgenes_andcounts, metadata, c("Tol" = "blue", "Rej" = "green"))
}
