% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goGrouping.R
\name{genelistFromGOGroup}
\alias{genelistFromGOGroup}
\title{Gene List from Go Grouping}
\usage{
genelistFromGOGroup(gogroup, rownum = NULL, goid = NULL)
}
\arguments{
\item{gogroup}{dataframe, output of goGrouping function}

\item{rownum}{number, which row to return genes from}
}
\value{
dataframe, MGI symbols
}
\description{
Gene List from Go Grouping
}
\examples{
genelistFromGOGroup(goGrouping(one_vs_two), 2)
}
