% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPCAobj.R
\name{get_pca_obj}
\alias{get_pca_obj}
\title{Get PCA object}
\usage{
get_pca_obj(norm_df, metadat, varianceRemove = 0.05)
}
\arguments{
\item{norm_df}{dataframe of normalized counts (rlog df)}

\item{metadat}{metadata dataframe}

\item{varianceRemove}{Remove this % of variables based on low variance.}
}
\value{
pca_obj
}
\description{
Get PCA object
}
\examples{
get_pca_obj(rlog_df, meta, 0.01)
}
