% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedGenes.R
\name{get_shared_genes}
\alias{get_shared_genes}
\title{Get Shared Genes}
\usage{
get_shared_genes(suffix_list, dataframe1, dataframe2, ...)
}
\arguments{
\item{suffix_list}{list of strings, one for each data frame inputted, in the same order}

\item{dataframe1}{data frame, contains at least MGI symbol and description, log 2 fold change, adjusted p value, and change direction}

\item{dataframe2}{data frame, contains at least MGI symbol and description, log 2 fold change, adjusted p value, and change direction}

\item{...}{data frames, contains at least MGI symbol and description, log 2 fold change, adjusted p value, and change direction}
}
\value{
data frame, contains all the genes significantly up/down in the same direction across given data frames including fc, p val
}
\description{
Get Shared Genes
}
\examples{
get_shared_genes(c(".naive_tol", ".naive_rej"), sig_NaiveTol, sig_NaiveRej)
get_shared_genes(c(".naive_tol", ".naive_rej", ".naive_ex-tol"), sig_NaiveTol, sig_NaiveRej, sig_NaiveExTol)
get_shared_genes(c(".naive_tol", ".naive_rej", ".naive_ex-tol", ".naive_memory"), sig_NaiveTol, sig_NaiveRej, sig_NaiveExTol, sig_NaiveMemory)
}
