% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGSEA.R
\name{go_GSEA}
\alias{go_GSEA}
\title{GSEA Gene ONtology}
\usage{
go_GSEA(degs, ontol = "ALL", pval_cutoff = 0.05)
}
\arguments{
\item{degs}{annotated, pairwise comparison dataframe, *not* the significant genes}

\item{ontol}{string, either "BP", "MF", "CC", or "ALL", defaults to ALL}

\item{pval_cutoff}{number, defaults to 0.05, p value cutoff value}
}
\value{
GSEA object
}
\description{
GSEA Gene ONtology
}
\examples{
go_GSEA(tol_vs_naive, ontol = "BP")
}
