% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goGrouping.R
\name{goGrouping_all_levels}
\alias{goGrouping_all_levels}
\title{Get Go Groups from all levels (2-6)}
\usage{
goGrouping_all_levels(
  genes,
  pval_cutoff = 0.05,
  l2fc_cutoff = 0.5,
  ontol = c("BP", "MF", "CC", "ALL")
)
}
\arguments{
\item{genes}{dataframe, list of DEGs, must be annotated and have Log2FoldChange and Adj_P_Value columns}

\item{pval_cutoff}{number, defaults to 0.05}

\item{l2fc_cutoff}{number, defaults to 0.5}

\item{ontol}{string, "BP", "CC", "MF" or "ALL"}
}
\value{
gg, dataframe
}
\description{
Get Go Groups from all levels (2-6)
}
\examples{
goGrouping_all_levels(one_vs_two, 0.01, 0.8, "BP")
}
