% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goGrouping.R
\name{goGrouping_rlog_all_levels}
\alias{goGrouping_rlog_all_levels}
\title{Get Go Groups from all levels (2-6) from rLogDF}
\usage{
goGrouping_rlog_all_levels(rlogdf, ontol = c("BP", "MF", "CC", "ALL", "BP_MF"))
}
\arguments{
\item{rlogdf}{rlog dataframe - ideally filtered to only include significantly changed/relevant genes}

\item{ontol}{string, "BP", "CC",  "MF", "ALL", or "BP_MF"}
}
\value{
gg, dataframe
}
\description{
Get Go Groups from all levels (2-6) from rLogDF
}
\examples{
goGrouping_rlog_all_levels(one_vs_two, 0.01, 0.8, "BP")
}
