% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennDiagram.R
\name{plot_venndiagram}
\alias{plot_venndiagram}
\title{Plot Venn Diagram}
\usage{
plot_venndiagram(degs_list, plt_title = "", edgecolor = c("#000000"))
}
\arguments{
\item{degs_list}{list of named lists, lists of MGI symbols from significant DEGs of pairwise comparisons, should be named with the label}

\item{plt_title}{string, name of the plot}

\item{edgecolor}{list of strings, colors for the circles of the venn diagram, should be in the same order as the lists}
}
\value{
ggplot venn diagram
}
\description{
Plot a venn diagram showing the number of shared DEGs between 2+ comparisons
Can be any number of lists but for best Venn Diagram
}
\examples{
plot_venndiagram(list("Treg Depleted" = ranked_TregNaive$MGI_Symbol, "Tolerant" = ranked_TolNaive$MGI_Symbol))


}
